revert arrondi distances

+++ src/main/java/model/Instance.java
@@ -17,7 +17,7 @@ public class Instance {
     private final List<Customer> customers;
     private final Fleet firstEchelonFleet;
     private final Fleet secondEchelonFleet;
-    private final List<List<Long>> distanceMatrix;
+    private final List<List<Double>> distanceMatrix;
 
     public Instance(List<Depot> depots, List<Satellite> satellites, List<Customer> customers, Fleet firstEchelonFleet, Fleet secondEchlonFleet) {
         this.depots = depots;
@@ -32,7 +32,7 @@ public class Instance {
                                 .flatMap(x -> x.stream())
                                 .collect(Collectors.toList());
         for (Site siteLigne : sites) {
-            ArrayList<Long> distances = new ArrayList<>();
+            ArrayList<Double> distances = new ArrayList<>();
             for(Site siteColonne : sites){
                 distances.add(siteLigne.computeDistance(siteColonne));
             }
@@ -62,7 +62,7 @@ public class Instance {
         return secondEchelonFleet;
     }
 
-    public List<List<Long>> getDistanceMatrix() {
+    public List<List<Double>> getDistanceMatrix() {
         return distanceMatrix;
     }
